Converts life to double

+++ src/objects/entity.cpp
@@ -1,6 +1,6 @@
 #include "entity.hpp"
 
-Entity::Entity(double size, int life) : 
+Entity::Entity(double size, double life) : 
     sf::CircleShape(size),
     life(life),
     total_life(life)
@@ -12,7 +12,7 @@ bool Entity::alive()
     return life > 0;
 }
 
-void Entity::damage(int n)
+void Entity::damage(double n)
 {
     if (life > 0)
     {
@@ -20,7 +20,7 @@ void Entity::damage(int n)
     }
 }
 
-void Entity::regen(int n)
+void Entity::regen(double n)
 {
     if (life < total_life)
     {
@@ -30,7 +30,7 @@ void Entity::regen(int n)
 
 double Entity::lifePercent()
 {
-    return (double)life / (double)total_life;
+    return life / total_life;
 }
 
 void Entity::update()
